The great builtin migration of 2019
Stage 1:
This moves all tested constructors to the builtin trampolines and additionally
fixes `type.__call__` EX for importlib. Last, moves implementation of
`str.__new__` into managed code.

Additionally, it migrates some other methods (like those of `property`,
`classmethod`, `staticmethod`) because they are easy.

Stage 2:
Reorganize structure of builtin methods

Eliminates builtinTrampolineWrapper (which curried the native entrypoint
pointer into the function object) in favor of storing the native entrypoint
pointer in the function.code.code and the trampolines (which know how to unpack
that) in the function object's entrypoints.

This diff introduces NativeMethods (nee BuiltinMethods) and BuiltinMethods.
After this diff, we should only ever be adding to the BuiltinMethods array, and
removing from the NativeMethods array.

So, in short:
OLD:
* function.entry* is Wrapper<entry>
* function.code.code is NoneType -> wrapped builtin

NEW:
* function.entry* is Trampoline (no wrapper)
* function.code.code is SmallInt -> builtin address

Based on Facebook D14091607