Fix unmarshalling of some negative large ints
The condition for whether we needed to append an extra sign extension digit to
a large int did not take into account whether a carry did or did not happen
from lower digits.

This changes the code to check for sign extension after negation when the
carries are already applied.

Based on Facebook D14069538