Create helper to convert ByteArray to Str
Implement the helper, switching on the array size to determine whether the result is SmallStr or LargeStr, and copying the array into the new string.

This will be used by `__repr__` and `decode` functions to allow strings to be built one byte at a time.

The first use of this helper will be in D14055779 for `bytes.__repr__`.

Based on Facebook D14112069