Implement the Struct Sequence C-API
**This implements**:
* `PyStructSequence_GetItem` && `PyStructSequence_GET_ITEM`
* `PyStructSequence_SetItem` && `PyStructSequence_SET_ITEM`
* `PyStructSequence_New`
* Modifications to `PyStructSequence_NewType`

**Background**:
Struct Sequences are a special kind of type in CPython. They have the following features:
* Subclass of tuple so they have index accessing to the visible fields.
* Fields names that map to the tuple values.
* Hidden fields that are accessible by name or with the index through the C-API
Example:
```
# Foo is a struct sequence type
class Foo(tuple):
  foo = index at 1
  bar = index at 2
  baz = hidden index at 3

a = Foo((1,2,3))
a.foo
>>> 1
a[1]
>>> 2
a[2]
>>> Index out of range
a.baz
>>> 3

// C-API
PyObject* foo = PyStructSequence_GetItem(a, 0);
PyStructSequence_Settem(a, 1);
PyObject* baz = PyStructSequence_GetItem(a, 2);
```

**Design**:
To achieve the same behavior, the following design was followed:
* Internal Tuple: The Struct sequence being a sub-type of tuple will have a tuple of exactly the size of the number of visible fields.
* Hidden Fields: The rest of the fields will be stored as instance attributes.
* Named Members: These will be descriptors that will either access the tuple through index or the instance attributes.

**Attribute Access**:
In order to achieve the mentioned behavior, a descriptor is constructed (_structseq_fields) with the following properties:
* `__get__`: Check if the index is within the tuple bound. Otherwise, look for the name in the instance attributes.
* `__set__`: Always throw, this will make the type immutable
* `GetItem`: Call back to Python, same behavior as `__get__`
* `SetItem`: Check if the index is within the tuple bound and insert there. Otherwise, get the field name and insert it into the instance.

Please note that this implementation has the downside that it bypasses the normal getattr and setattr in order to achieve the same compatibility. This means that calls to these types will not be cached and we will have to always walk the attribute values to find it.

**Length**:
This defaults to the length of the tuple backing the struct sequence. Therefore, printing out the visible size.

**Iter**:
This default to the iter from tuple, therefore, only iterating over the visible members

**Repr**:
This currently does not print out the member names as we don't have support to iterate through members yet. However, it does print the values in a tuple-like format.

**Unnamed Fields**:
Our implementation will **ban** the use of unnamed fields in struct sequences. An ad-hoc investigation showed that posixmodule is the only module that is using unnamed fields as they were introduced to support a specific posix use case: https://fburl.com/bo9x1ylm.

That being said, we'll rely on what the implementation of PyPy is doing - which is banning the use of unnamed fields: https://fburl.com/b1rha294.

By banning the use of unnamed fields, the implementation of struct sequences become much simpler. For instance, we can guarantee that the size of the tuple backing the struct will always be the size of the visible length. Also, we can guarantee that the remaining fields will have a name and can be set into the instance's attributes.

**Compatibility**:
The current design should allow for almost full compatibility with CPython. There are a couple of exceptions:
* *PyStructSequence_GetItem*: If this C-API is called before setting a value to the field, CPython will return `nullptr` while we return `Py_None`. Please note that if the field is accessed by name with a C-API such as `PyObject_GetAttrString`, then both do return `Py_None`.
* *PyStructSequence_SetItem*: This only does a type check to see if the passed in object is an instance of tuple. However, there is no way of guaranteeing that this is indeed a struct sequence. CPython doesn't do any type checks either. So while we are keeping compatibility with CPython this doesn't protect us from bad user code calling this function without a struct sequence.

Based on Facebook D13758236