Fix returning from inside a generator
The old way of directly translating `return foo` into `raise
StopIteration(foo)` meant that an `except:` block could swallow the
return. Match what CPython does, and defer raising the exception until after
the generator is finished.

While I was in here, I also made sure we raise a StopIteration when you try to
resume a finished generator, whether it was finished from a return or a raise.

Finally, a couple tangentially-related bits of cleanup: fix the type of an
exception raised in `GeneratorBuiltins::dunderNext` and make `typeName()` in
`test-utils.cpp` resilient to being given an `Error`.

Based on Facebook D13945950