Add importlib to the runtime
This fully freezes importlib into the runtime to use CPython's import mechanism out of the box.

In order to make this work during initialization, we execute `_bootstrap` and `_bootstrap_external` and run `_bootstrap._install` which sets into the runtime `importlib` as the implementation of `import`.

Then when importing modules, this executes `_bootstrap._find_and_load(name)` which is in charge of looking into `sys.path` to find a matching `.py`, `.pyc`, or `.so`.

Please note that only `.pyc` work for now which is the simplest scenario.

Based on Facebook D14040945