Make {list,byteArray}EnsureCapacity more intuitive
The existing implementation of `listEnsureCapacity` and `byteArrayEnsureCapacity` called for an index and guaranteed a safe `atPut` at that index.

That design is counter-intuitive, hence several callsites to `listEnsureCapacity` passed a desired minimum capacity instead.

This diff changes both functions to take the desired capacity instead, factoring out a common growth method.

Based on Facebook D14184128