Implement Unicode{Decode|Encode|Translate}Error builtins
Adds `UnicodeError`'s getters and setters, along with its three subclasses' builtins.

Since all three of the error functions have the same superclass and have the same fields and initializers (with the exception of `UnicodeTranslateError`, whose `__init__` does not take an `encoding` or set its `encoding` field), most of this diff is the exact same functionality tested on three different classes.

Based on Facebook D13927547