Implement _PyUnicode_Ready
`_PyUnicode_Ready` in CPython makes sure that strings are in their "canonical form", and makes any appropriate conversions between string kinds.

Since we only have a single kind and form, and since we have no concept of a "non-ready" string, this function should just be a no-op that always returns 0.

Based on Facebook D14266296