Add code point interfaces to Str
This adds a basic algorithms for UTF-8 to code point decoding and
converting a code point index into a UTF-8 code unit index.  It also
adds a fast algorithm for counting the number of code points.

String data in our runtime is guaranteed to be structurally-valid
UTF-8.  That simplifies these algorithms as they do not need to
anticipate invalid data.  These methods are intended to be simple and
fast and not to be generalized to handling malformed data.

A future change will integrate these interfaces with the str iterator
and the str dunder getitem.

Based on Facebook D14277247