Implement PyNumber_InPlace C-API functions
This diff implements the remaining C-API functions in the number protocol that use the operator module.

This includes PyNumber_Power with two arguments and all of the in-place operations.

The only remaining operations are PyNumber_Power and PyNumber_InPlacePower, both in the case with three arguments.

Based on Facebook D14309829