Implement extended dumpers
When dumping things like Code or Function it is often useful to see actual details of these objects. While on the other hand when code or function is part of deeper structure and not the toplevel thing you dump it is often distracting to see the details. This introduces `dumpExtendedXXX` functions that are used as the toplevel dump functions that dump some object types in extended detail, while `operator<<` keeps dumping the shorter representations.

This implements a dumper for code and function objects including a very simple bytecode dumper.

Based on Facebook D14236891