Str::codePointIndex => Str::offsetByCodePoints
Summary: Modifies the existing method to get the internal code unit index in a string from a code point index. Adds a start index to count from, makes it clearer that the index refers to code units, and calls the other input count so it does not get confused with indexes, which should refer to code units internally.

Based on Facebook D14368515