Add bytes.decode and bytes.encode and the _codecs module
Adds the `_codecs` builtin module along with `_codecs.decode` and `_codecs.encode`, both of which are slightly different from CPython's implementations, and closer to their bytes builtins version in that they will try to fast-pass basic codecs to their respective codecs.

Making everything dispatch to `_codecs` is done so as to have a single source of truth codec, and fast-passing is necessary for common use-cases.

Based on Facebook D14377262