Fix function.__get__ on type(None)
Pythons `__get__` is used to lookup descriptors on instances and on
types. To differentiate the two cases, the 2nd parameter (`obj`) is set
to `None` when a descriptor is looked up on a type.
This leaves us with an ambiguity: Looking up a descriptor on
the `None` object also has the 2nd parameter as `None`: `None.__repr__`
and `type(None).__repr__` will both result in a call to
`function.__get__(self, None, type(None))`.

cpython can work around this problem by differentiating between `None`
and `NULL` on the 2nd parameter when calling `tp_descr_get`, this does
not work in pyro however because we do not have the `tp_descr_get` layer
and cannot pass NULL. We also cannot change how `__get__` works
because it is python API.

This only leaves us with special casing the situation when `None` or
`type(None)` is involved:
- Special case lookup of attributes on `type(None)` to immediately return the
  plain attribute and skip the `__get__` mechanism.
- Special case `function.__get__` to always return a bound method when
  the type parameter is `type(None)`.

Based on Facebook D14378331