Clean up use of UnboundValue in BaseException attributes
cause, context, and traceback all start as UnboundValue, so we can
return NULL (rather than None) to C-API clients. No other callers care about
the difference between None and UnboundValue, so change the normal accessors to
fold UnboundValue into None and add raw accessors that return it unchanged.

Things are still sub-ideal even with this fix: if managed code asks for one of
these attributes, we'll pass the UnboundValue back unchanged. I'd like to
resolve that at a later date; this is just a quick fix to make an upcoming diff
less gross.

Based on Facebook D14447384