Implement PyFile_Write{String,Object}
As with all stream-related things, we only currently support the
integers 1 or 2, for stdout and stderr. Also remove printStr() and call
fileWriteObject() instead.

One "quirk" of PyFile_WriteString() is that when you call it with a pending
exception, it returns -1 and prints nothing. This is to allow repeated calls to
it without checking each one for success. I preserved that behavior in
PyFile_WriteString() itself but not our internal version, fileWriteString().

Based on Facebook D14427627