Introduce memoryview object, implement memoryview.__getitem__
Add RawMemoryView object (aka memoryview in python). This is a descriptor for a block of memory.
The initial version holds a pointer to a bytes object, a format string to describe the type of the contained data and a read-only flag. There is no support yet for multidimensional data shapes, strides, ...

- Introduce `RawMemoryView` class, layout id, etc.
- Add the memoryview-builtins infrastructure.
- Implement `memoryview.__getitem__`.
- Introduce memoryview testing helpers.

Based on Facebook D14469746