Add sys.maxsize
Implement `sys.maxsize` as `kMaxWord`.
Python defines it as "the largest supported length of containers." which would be `RawSmallInt::kMaxSize` in pyro. However when you look at users in the standard library, it is typically used to figure out the size of a machine word (to differentiate between 32 and 64bit architectures) or as a marker for an invalid index. These uses are better served by the `kMaxWord` value.

Based on Facebook D14601382