Implement PyErr_Display()
This performs the meat of the work done by PyErr_Print() (which is in
the next diff). CPython's implementation of the printing code ignores
exceptions from PyFile_WriteString() in many paths and attempts to continue
printing (much) later parts of the exception chain. I decided to not match this
and just treat exceptions normally, with the exception of swallowing them at
the top to maintain high-level compatibility with CPython. This makes for
cleaner, easier-to-follow code, and I don't think it will make a difference in
any practical situations, but I'm open to being convinced otherwise.

Rename BaseException.cause to the correct `__cause__`, and add
`BaseException.__suppress_context__` for use in the printing code. Add
Runtime::attributeAtId() for convenience.

Based on Facebook D14427630