Change tagging for existing immediates
We will be changing the tagging scheme for RawObject over several diffs.

The existing scheme utilized a 2-bit low tag and a 5-bit high tag.
```
*** *0 - SmallInt
*** 01 - HeapObject
**0 11 - Header
001 11 - Bool
011 11 -
101 11 - Error
111 11 - SmallStr
```

The new scheme maintains a 5-bit high tag, but we have increased the low tag to 3-bits to make more space for `SmallBytes`, `NotImplemented`, and `Unbound` immediates.
```
** **0 - SmallInt
** 001 - HeapObject
** 011 - Header
00 101 - SmallBytes
01 101 - SmallStr
10 101 - Error
11 101 - (unused)
00 111 - Bool
01 111 - NotImplemented
10 111 - Unbound
11 111 - NoneType
```

This diff changes the `kTag`, `kTagMask`, and `LayoutId` for each of the existing immediates and rearranges their classes and methods into the new order.

It also changes `NoneType::object()` to `kMaxUword` so that we can `memset(foo, -1, bar)` to fill an object with `None`.

The introduction of immediate `SmallBytes`, `NotImplemented`, and `Unbound` has been left for later diffs.

Based on Facebook D14582473