Implement codecs.ascii_decode
Implement the logic for decoding ASCII according to CPython's implementation.

Since this function is supposed to be the same as _codecs.ascii_decode, it has the same return type of a tuple whose first element is the decoded string and whose second element is the length of the encoded bytes passed in.

Also introduces the generalized decoder error handler. Each decoder calls this error handler, which creates a UnicodeDecodeError, passes it to an errorhandler, writes the function's output to a passed in ByteArray, and returns the potentially changed bytes object along with the new position at which to continue decoding. The function's full documentation can be found in its docstring

Also fixed `ignore_errors` to properly return 0 as the tuple's second element

Based on Facebook D14378075