Import atexit module
Import the atexit function along with an implementation of `_Py_PyAtExit` which registers a C function into the runtime which gets executed at destruction time.

This also adds module finalization in `freeApiHandle` to correctly call the free function on a module once the runtime is being destroyed.

Based on Facebook D14758147