Don't preemptively add ___extension___ attribute to PyType_FromSpec() layouts
I think this was an attempt to give the type an in-object attribute,
but that's not how Layouts work. We don't currently have a convenient way to do
this, so just let the let the Layout transition happen the first time an
`___extension___` attribute is set.

If you created an instance of a type made with `PyType_FromSpec()`, this would
cause `ApiHandle::{new,borrowed}Reference()` to abort while checking for the
`___extension___` attribute.

Based on Facebook D14763496