Increase initial ensured array capacity
After the introduction of `SmallBytes`, we would like `ByteArray` to be backed by either the empty bytes singleton or a `LargeBytes`.

We accomplish this by ensuring that the length of underlying array is at least `kWordSize`.

This will future-proof us for the introduction of `StrArray` as well.

Based on Facebook D14748215