Move bytes creation to managed code
Let managed code handle iteration over an object to create a `bytes` from an iterable that does not have a special case.

Push `bytesFromTuple` into the runtime both so we can avoid the initial `std::memset` and so we are future-proof for `SmallBytes` later.

Add an optimization to `Runtime::bytesRepeat` to delegate to `Runtime::bytesNew` when `length == 1`.

Based on Facebook D14809921