Make Bytes a wrapper around LargeBytes
The existing `RawBytes` class becomes `RawLargeBytes`, and we introduce a new `RawBytes` as a wrapper that delegates to `RawLargeBytes`.

Because we want callers (except `Runtime` and `ByteArray`) to go through `Bytes`, make `LargeBytes` methods private. Move the resizing in `PyBytes_Resize` from the C-API to the runtime for fewer memory writes.

Introduction of `SmallBytes` has been left for a later diff.

Based on Facebook D14618896