Add macros for warn_unused(_result)? attributes, use in a couple places
Add `NODISCARD` and `WARN_UNUSED` macros for the `warn_unused_result`
and `warn_unused` attributes, respectively. Ideally, we'll use `NODISCARD` (or
`[[nodiscard]]` once we're on C++17) on every `RawObject` return value that
could be an `Error`. For now, I'd like to add the macro and start the
convention that we add it to any new functions, and convert existing functions
as we have time. I also converted the existing uses of
`__attribute__((warn_unused))` to `WARN_UNUSED` for consistency.

Based on Facebook D14857708