Implement PyUnicode_DecodeLocale{AndSize} and _Py_normalize_encoding
Implement `PyUnicode_DecodeLocale{AndSize}` similarly to CPython, but without some of their optimizations. I added a task to eventually follow their optimizations if it ends up becoming an issue. Since it can all be done using the C-API, and since this shouldn't ever be a performance bottleneck, I copied most of the implementation from CPython.

To effectively implement it, we need to use `Py_DecodeLocale`, which requires `_Py_normalize_encoding`, so I just copied CPython's implementation of it exactly.

We also need `PyCodec_StrictErrors`, which we've already implemented and tested in native code, so I added a call to `_codecs.strict_errors`.

I've also gone ahead and fixed `newStrFromUTF32` so that I can just follow CPython and call `PyUnicode_FromWideChar`.

Based on Facebook D14689587