Run Python tests with Pyro and CPython
Modifies the target determinator to run python tests with both pyro and cpython.

**Writing a Test**:
Python tests should be suffixed with `_test` and they should be in the same directory as the module they are testing. This matches the current way we test C++ files in the runtime.

Given that this also moves `pyro/library/tests/test_unittest.py` to `pyro/library/unittest_test.py`

**Running tests with CPython in Diffs**:
If CPython is executed with the tests as is, it will use the modules in the `library` directory (instead of `cpython/Lib`). To avoid that, this copies the tests into the build directory (`build/tests`) and then runs CPython over them.

That this will only copy tests that are prefixed with an alphabet character (i.e. a-z A-Z). Those modules prefixed with an underscore are exclusively internal to our runtime, so they won't be copied over. This allows us to start writing higher level tests for internal/helper modules that do not exist in CPython.

Note that there are currently two exceptions which have to be fixed:
* `builtins.py`: This has to be renamed to `_builtins.py` to match the convention.
* `_collections_abc.py`: This should be tested through `collections` which is the public API (i.e. `collections._collections_abc`).

Furthermore, this will be running more than one set of unit test files. Therefore, this uses `xargs` to run a python process for each of the files. If any of those fails, xargs will return an exit code > 0 making the sandcastle job fail.

**Running tests locally**:
The build directory will now contain two new scripts:
* python-tests-pyro.sh: Runs all the python tests found in library/*_test.py using Pyro
* python-tests-cpython.sh: Runs all the python tests found in library/[a-zA-Z]*_test.py using CPython

Simply, build the project and run the script to run all python tests no need to rebuild. If only one test is required to run, pass the name of the file as an argument to the script. I.e. `./python-tests-pyro.sh time_test.py`

Based on Facebook D14875197