Add missing int subclass checks
Currently, the pattern is to check if an object `isInstanceOfInt` and then pass that into a handle. However, this is correct as subclasses of int can't be placed into an `Int` handle.

Therefore, this change tries to consolidate the int checks throughout the codebase and mark all the places that need to have the int properly pulled out from a subclass.

Based on Facebook D14814124