Include proper filenames in compiled code objects
Force all callers of `Runtime::compile*` to provide a filename, which
is passed to the bytecode compiler. Callers that really don't have a filename
still pass a dummy value, but this gets the important callsites and changes the
traceback on a crash I was debugging from {P62405020} to {P62404980}.

I also simplified the flow in `main()` a bit while I was in there.

Based on Facebook D14904650