Add simple RawObject lldb printer
I kept it very basic for now, with the idea that you can fall back to
the more complete functions in `runtime/debugging.cpp` if you need more
information about the value. This has a couple advantages over `debugging.cpp`:
- It applies anywhere a `RawObject` is printed, without you having to do
  anything (like `fr var`, or return values after running `finish`).
- It doesn't require a live process, so it will work on a core file. This
  doesn't matter much now since all of our tests are small, but it will be very
  helpful once we're dealing with cores from prod.

This does duplicate knowledge about type layouts that until now had one
canonical home in C++. I expect these things will be pretty stable, but we can
replace it with codegen if/when that becomes an issue.

Based on Facebook D14854893