Add operator slot support to PyType_FromSpec()
These are referred to as operators within CPython, but it's really any
PyType_Slots that have a corresponding managed method name. For example, if you
provide a `Py_tp_hash` slot to `PyType_FromSpec()`, we will now synthesize a
`__hash__` method on your type that fusses with the arguments and return value
as needed. There are three main parts to this:
- `typeobject.cpp:kSlotdefs` contains the mapping from slot ID to name and wrapper
  function. This was copied from CPython and massaged into types that we
  understand.
- The additions to `trampolines.cpp`: These all perform subsets of the work
  done by `builtinTrampoline{,Ex,Kw}`, but they're so drastically simpler that
  I decided it was worth keeping them separate for now.
- The wrapper functions themselves, back in `typeobject.cpp`, which do the
  argument/return value handling. Unfortunately it's not as simple as "convert
  n arguments and 1 return value", since some of the C function pointers take
  or return non-`PyObject*` values, and some of the wrappers have extra
  logic. If you want to compare these with their CPython counterparts, look in
  `typeobject.c`: `wrapInit()` corresponds to `wrap_init()`, etc...

The trampolines and wrappers I added subsumed the duties of
`extensionTrampoline` and friends, so those have been deleted. This exposed a
bug in the old code: `extensionTrampoline` was leaking a reference to the
`PyObject*` return value from the slot function by not decrefing it after
converting to a `RawObject`. I spent a couple of days trying to fix the
underlying issue, but got caught up in some larger GC issues. I filed T42827325
and kept the reference counting bug for now; reading that task should help you
understand the related TODO comments in this diff.

Based on Facebook D14761955