Add sys._getframe_{code,globals,lineno}
This diff adds functions that produce similar functionality to `sys._getframe` while not
fully implementing the frame inflation and traversal logic. We use it to implement warnings,
and can be used by any feature that needs frame info but not the frame object (which implies
getting the full stack trace).

Based on Facebook D14938575