Implement gcmodule
This change adds gcmodule.cpp that just tracks objects allocated via GC_Malloc. CPython's type PyGC_Head is replaced with a simpler ListEntry type since ListEntry is confirmed to be internal to this module.

Note that gcmodule is unrelated to Pyro's tracing GC, but we need to implement this module to simulate its side effects of calling tp_traverse of tracked objects when PyGC_Collect() is called.

Track/untrack operations for ListEntry are added to Runtime in case it's reused for other internal tracking purposes.

Based on Facebook D14898959