Implement codecs.utf_8_encode and _?PyUnicode_AsUTF8String
Follow the same architecture as `codecs.ascii_decode`, except as an encoder

Also introduces the generalized encoder error handler. Each encoder calls this error handler, which creates a UnicodeEncodeError, passes it to an errorhandler, and returns the errorhandler's result. The function's full documentation can be found in its docstring

Since this encoder switches between managed and native code in the slow path of a non-standard error handler, it needs to properly translate between codeunit indices and codepoint indices in native code when reading and returning to managed code.

Based on Facebook D14378147