Set errno to 0 before calling strtoll
Set `errno` to 0 in `intFromBytes` before calling `strtoll`, since according to its spec (https://linux.die.net/man/3/strtol), the caller is reponsible for clearing it.

This was causing `BuiltinsModuleTest.IntFromBytesWithZeroBaseReturnsCodeLiteral` to fail locally because it was picking up that `errno` was set to `EINVAL` by a previous test.

Based on Facebook D15049548