Make tests more consistent
Increase consistency of some tests. This uses the following pattern and puts it at the beginning of the test:
```
Runtime runtime;
Thread* thread = Thread::current();
HandleScope scope(thread);
```
Change any other `Thread::current()` usage in the test to `thread`.

This changes the tests in `thread-test.cpp` and `interpreter-test.cpp` which I plan to work on in other diffs.

Based on Facebook D15056426