Fix builtinBases to depend on kAttributes
Builtin types access their data at fixed offsets. This is incompatible with multiple inheritance when two builtin base classes expect different data at the same offset.

We solve this by only allowing multiple inheritance when the builtin bases are on the same path in the inheritance hierarchy.
- Manually mark low-level builtin-types to be their own builtin base: `int`, `bytes`, `str`, `float`, `complex`, `object`, etc.
- In case of object subclasses (such as exception types) mark the type as its own builtin base if it defines additional attributes, otherwise inherit the builtin base from the supertype.

Note: This is the same as the "solid base" concept in cpython.

An upcoming diff will change `computeBuiltinBase` to actually properly compute a class's builtinBase from these newly-fixed bases.

Based on Facebook D15055595