Implement weakref.proxy
weakref.proxy(o, callback) returns an object that behaves like o, but unhashable, and callback is called when all references are gone.

Note that the behavior of weak ref looks different in pyro, which is affecting the proxy's behavior that depends on ref. See T43270097 for details. Left a TODO to fix the issue aftewards.

Based on Facebook D14989849