Implement `codecs.utf_16_encode` and `_?PyUnicode_EncodeUTF16`
Implement a utf-16 encoder in the same way as the other encoders.

Since utf-16 is a two-byte encoding, endianness starts to come into play. CPython supports outputing in either the native endianness or the given endianness. When a specific endianness is provided, only the encoded bytes are outputted. However, when native endianness is used, a BOM mark (\xFEFF) is written to the beginning of the bytes object to signify the endianness used. Since we have control of both encoding and decoding, we can just use little endian as our "native" endianness, and expect any user who wants something different to pass in their desired endianness.

Also implements `utf_16_le_encoding` and `utf_16_be_encoding` by passing in a byteorder int to the generalized utf-16 encoder

Based on Facebook D14923228