Add a kind to Error objects
This adds a basic tag to Error objects, letting us distinguish between
a few different types of error. `Error::error()` still exists as a generic
Error, but it only has a few uses left. All other callsites to
`Error::object()` have been converted to one of the more specific kinds.

I also audited every call to `thread->hasPendingException()` to see if it could
be replaced with an error kind check, and about half of them were (most of the
holdouts are either in C-API code where we have to disambiguate a nullptr
return value, or in test code that's being thorough).

I didn't audit calls to `obj.isError()`, partially because there are over 1400
of them, but also because they're not doing any harm: `isError()` checks for
any kind of error, so we can clean things up as needed/incrementally.

Based on Facebook D15156009