Implement issubclass and type.__subclasscheck__
Implement `issubclass` with similar code paths as CPython, but put the fast checks in native code, including `_type_checkexact` and `_type_issubclass`.

Implement `type.__subclasscheck__` using `_issubclass` with the arguments flipped.

We are still missing some attributes on `type` to use an arbitrary `__subclasscheck__`, so leave that as a TODO.

Based on Facebook D15036121