Finish codecs.encode and codecs.decode
Since `lookup` is already implemented, we can remove the `TODO`s in `codecs.encode` and `codecs.decode`.

Make sure that we follow the same error checking as CPython. Only difference is that CPython wraps any error caused during the execution of the encoder/decoder with an error of the same type with a different message. Since Pyro isn't going for full error message parity, there's no reason to go through the complexity of figuring out which error types are possible to wrap around, along with setting each of the appropriate attributes when the expected error type is the same.

Based on Facebook D15283705