Avoid default constructor of HandleScope for dict functions
This change passes `Thread* thread` pointer to dict related runtime functions so that `HandleScope` can be created with the given `thread` without calling Thread::current() in it. This is aligned with our long term goal to pass `thread` to runtime functions.

In this diff, the aforementioned change is made only to `Runtime::*dict*` functions in `runtime.h` and all other changes in this diff are for their callers.

This change makes a minor improvement in terms of the callgrind instruction count (1.1%), but its CPU time improvement is not confirmed due to the variance in the devvm where benchmark was run (`python-control` is the one compiled with this change):

```
[djang@devvm3983.prn3 ~/fbsource/fbcode/pyro] ./util/measure_command.py --callgrind --time ./build/python-control ./benchmarks/richards.py  && ./util/measure_command.py --callgrind --time ./build/python-base ./benchmarks/richards.py
{
  "instructions": 51195323312,
  "run_time_secs": 9.601418562233448
}
{
  "instructions": 51767097807,
  "run_time_secs": 9.977837000973523
}
```

Based on Facebook D15275601