Implement C-APIs for {static,class,}method
This uses the existing staticmethod and classmethods to implement `PyStaticMethod_New` and `PyClassMethod_New`. Furthermore, `PyMethod_New` was a natural extension of these C-API by allowing any arbitrary object to be bounded to the function.

This also uncovered a bug in `Interpreter::callEx` where it assumed that any object passed through it was going to be either a Function, a Bound Function, or a Callable. This adds an extra check to raise an error if it's none of those.

Based on Facebook D15146338