Merge icFind and icUpdate
I originally split `icFind` and `icUpdate` into two operations, assuming we can use code paths that do not compute cache updates in cases of disabled or full caches.
It turned out that we rather want to always compute the value to put into the cache and `icFind` and `icUpdate` were only ever used adjacent to each other. We may just as well merge them now.
This is also fixes a bug where the code didn't check for negative indexes.

Based on Facebook D15280486