Clean up callable object preparation
`callEx()` was using an incomplete copy of the normalization logic in
`prepareCallableCall()` that didn't support non-Function descriptors in
`__call__`.
- Move the normalization code into `prepareCallable()`, which operates on
  `Object*`s rather than the stack.
- Move the stack normalization code from `callEx()` into a
  `prepareCallableEx()` helper.
- Use `prepareCallable()` in `prepareCallable{Call,Ex}()`.
- Update the error message to match CPython and fix relevant tests.

This may cause a few more stores than the old version of
`prepareCallableCall()`, but I think the increased simplicity is worth it,
especially since this is already a bit of a slow path. Any caching we add to
this path will likely sit in front of these functions anyway.

Based on Facebook D15282467