Replace dict.numEmptyItems for numUsableItems
Currently Dict tracks the number of empty items to decide whether to grow a dict or not, and dict grows when the number of empty items is less than 1/3 of the full capacity.

This way of tracking empty items is unintuitive since the number of empty items is *not* directly used for anything, but a derived metric used for growth decision.

This diff replaces it for the number of 'usable' items, set to 2/3 of the full capacity and decremented by 1 whenever an item is inserted into an empty item. This way, we can grow a dict when the usable item reaches zero.

Note that this change is for better maintainability rather than performance improvement.

The performance impact of this change slightly improved (callgrind: -0.23%, cputime: -2.83%) from a specific benchmark run but I believe it's within the measurement error range:

| base/control | callgrind | cpu time |
| -- | -- | --|
| base | 7166354568 | 1.22 |
| control | 7149460809 | 1.18 |

Based on Facebook D15332451